requirements.txt
tda-api==1.6.0
flask==2.2.3
selenium==4.8.3
pyopenssl
from-root
pandas-ta
Authlib~=1.2.0
pandas~=2.0.0
stopwatch.py
pytz~=2023.3